+++ magik-checks/src/test/java/nl/ramsolutions/sw/magik/checks/checks/CommentedCodeCheckTest.java
@@ -6,42 +6,125 @@ import java.util.List;
 import nl.ramsolutions.sw.magik.checks.MagikCheck;
 import nl.ramsolutions.sw.magik.checks.MagikIssue;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 /** Test CommentedCodeCheck. */
 class CommentedCodeCheckTest extends MagikCheckTestBase {
 
-  @Test
-  void testNoCommentedCode() {
-    final MagikCheck check = new CommentedCodeCheck();
-    final String code =
+  @ParameterizedTest
+  @ValueSource(
+      strings = {
         """
-        _method a.b
-            _local x << _self.call()
-            x +<< 10
-            write(x)
-            _return x
-        _endmethod""";
+      _method a.b
+          _local x << _self.call()
+          x +<< 10
+          write(x)
+          _return x
+      _endmethod
+      """,
+        """
+      _method a.b
+          # This is
+          # just a
+          # message,
+          # no real.
+          # code.
+      _endmethod
+      """,
+        """
+      _method a.b
+          a # z
+          b # y
+          c # x
+      _endmethod
+      """,
+        """
+      _method a.b
+          ## print(1)
+          ## print(2)
+          ## print(3)
+          ## print(4)
+          print(1)
+      _endmethod
+      """,
+        """
+      # Author         : Me
+      # Date written   : 01/95
+      # Date changed   :
+      """,
+        """
+      #
+      # Add/Remove
+      #
+      """,
+      })
+  void testValid(final String code) {
+    final MagikCheck check = new CommentedCodeCheck();
     final List<MagikIssue> issues = this.runCheck(code, check);
     assertThat(issues).isEmpty();
   }
 
-  @Test
-  void testCommentedCode() {
-    final MagikCheck check = new CommentedCodeCheck();
-    final String code =
+  @ParameterizedTest
+  @ValueSource(
+      strings = {
         """
         _method a.b
             #_local x << _self.call()
             #x +<< 10
             #write(x)
             #_return x
-        _endmethod""";
+        _endmethod
+        """,
+        """
+        #_method a.b
+        #_local x << _self.call()
+        #x +<< 10
+        #write(x)
+        #_return x
+        #_endmethod
+        """,
+        // Include the empty line in the commented code.
+        """
+        #_method a.b
+        #_local x << _self.call()
+        #x +<< 10
+        #x *<< 2
+        #
+        #write(x)
+        #x -<< 5
+        #_return x
+        #_endmethod
+        """,
+        // This will mark the indented commented body.
+        """
+        #_method a.b
+            #_local x << _self.call()
+            #x +<< 10
+            #write(x)
+            #_return x
+        #_endmethod
+        """,
+        """
+        #_method a.b
+            #_local x << _self.call()
+            #x +<< 10
+            #x *<< 2
+            #
+            #write(x)
+            #x -<< 5
+            #_return x
+        #_endmethod
+        """,
+      })
+  void testInvalid(final String code) {
+    final MagikCheck check = new CommentedCodeCheck();
     final List<MagikIssue> issues = this.runCheck(code, check);
     assertThat(issues).hasSize(1);
   }
 
   @Test
-  void testCommentedCodeTwice() {
+  void testInvalid2() {
     final MagikCheck check = new CommentedCodeCheck();
     final String code =
         """
@@ -58,77 +141,5 @@ class CommentedCodeCheckTest extends MagikCheckTestBase {
         _endmethod""";
     final List<MagikIssue> issues = this.runCheck(code, check);
     assertThat(issues).hasSize(2);
-  }
-
-  @Test
-  void testCommentedMessage() {
-    final MagikCheck check = new CommentedCodeCheck();
-    final String code =
-        """
-        _method a.b
-            # This is
-            # just a
-            # message,
-            # no real.
-            # code.
-        _endmethod""";
-    final List<MagikIssue> issues = this.runCheck(code, check);
-    assertThat(issues).isEmpty();
-  }
-
-  @Test
-  void testCommentsAfterCodeIgnored() {
-    final MagikCheck check = new CommentedCodeCheck();
-    final String code =
-        """
-        _method a.b
-            a # z
-            b # y
-            c # x
-        _endmethod""";
-    final List<MagikIssue> issues = this.runCheck(code, check);
-    assertThat(issues).isEmpty();
-  }
-
-  @Test
-  void testIgnoreMethodDoc() {
-    final CommentedCodeCheck check = new CommentedCodeCheck();
-    check.minLines = 2;
-    final String code =
-        """
-        _method a.b
-            ## Interesting
-            ## story.
-            # x
-            print(1)\
-        _endmethod""";
-    final List<MagikIssue> issues = this.runCheck(code, check);
-    assertThat(issues).isEmpty();
-  }
-
-  @Test
-  void testHeaderAccepted() {
-    final String code =
-        """
-          # Author         : Me
-          # Date written   : 01/95
-          # Date changed   :
-        """;
-    final CommentedCodeCheck check = new CommentedCodeCheck();
-    final List<MagikIssue> issues = this.runCheck(code, check);
-    assertThat(issues).isEmpty();
-  }
-
-  @Test
-  void testBlockWithEmptyLines() {
-    final String code =
-        """
-        #
-        # Add/Remove
-        #
-        """;
-    final CommentedCodeCheck check = new CommentedCodeCheck();
-    final List<MagikIssue> issues = this.runCheck(code, check);
-    assertThat(issues).isEmpty();
   }
 }